PR review updates #5615
