+++ ui/app/src/components/CreateSiteDialog/BlueprintForm.tsx
@@ -18,27 +18,16 @@ import React, { useEffect, useState } from 'react';
 import { makeStyles } from 'tss-react/mui';
 import Grid from '@mui/material/Grid';
 import TextField from '@mui/material/TextField';
-import FormControlLabel from '@mui/material/FormControlLabel';
 import GitForm from './GitForm';
 import { MarketplacePlugin, SiteState } from '../../models';
 import { defineMessages, useIntl } from 'react-intl';
 import PluginFormEngine from '../PluginFormBuilder';
 import { fetchAll } from '../../services/sites';
-import Switch from '@mui/material/Switch';
-import Typography from '@mui/material/Typography';
 
-const useStyles = makeStyles()((theme) => ({
+const useStyles = makeStyles()(() => ({
   form: {
     maxWidth: '600px',
     margin: '0 auto'
-  },
-  helpText: {
-    transition: `color .5s`,
-    display: 'block',
-    marginBottom: theme.spacing(2)
-  },
-  muted: {
-    color: theme.palette.text.secondary
   }
 }));
 
@@ -96,15 +85,6 @@ const messages = defineMessages({
   gitBranch: {
     id: 'createSiteDialog.gitBranch',
     defaultMessage: 'Git Branch'
-  },
-  createAsOrphan: {
-    id: 'createSiteDialog.createAsOrphan',
-    defaultMessage: 'Create the project from a remote repository as orphan (no git history)'
-  },
-  createAsOrphanHelpText: {
-    id: 'createSiteDialog.createAsOrphanHelpText',
-    defaultMessage:
-      'Creating the project as an orphan will dissociate the project from the source git repository and remove all history.'
   }
 });
 
@@ -286,20 +266,18 @@ function BlueprintForm(props: BlueprintFormProps) {
             helperText={formatMessage(messages.descriptionMaxLength, { maxLength: maxLength })}
           />
         </Grid>
-        {blueprint.id !== 'GIT' && (
-          <Grid item xs={12}>
-            <TextField
-              id="sandboxBranch"
-              name="gitBranch"
-              label={formatMessage(messages.gitBranch)}
-              fullWidth
-              onKeyPress={onKeyPress}
-              onChange={(event) => handleInputChange(event)}
-              placeholder={'master'}
-              value={inputs.gitBranch}
-            />
-          </Grid>
-        )}
+        <Grid item xs={12}>
+          <TextField
+            id="sandboxBranch"
+            name="gitBranch"
+            label={formatMessage(messages.gitBranch)}
+            fullWidth
+            onKeyPress={onKeyPress}
+            onChange={(event) => handleInputChange(event)}
+            placeholder="master"
+            value={inputs.gitBranch}
+          />
+        </Grid>
         {blueprint.parameters && (
           <PluginFormEngine
             parameters={blueprint.parameters}
@@ -316,28 +294,6 @@ function BlueprintForm(props: BlueprintFormProps) {
             handleInputChange={handleInputChange}
             onKeyPress={onKeyPress}
           />
-        )}
-        {blueprint.id === 'GIT' && (
-          <Grid item xs={12} sx={{ mb: 2 }}>
-            <FormControlLabel
-              control={
-                <Switch
-                  name="createAsOrphan"
-                  checked={inputs.createAsOrphan}
-                  onChange={(event) => handleInputChange(event)}
-                  color="primary"
-                />
-              }
-              label={formatMessage(messages.createAsOrphan)}
-            />
-            <Typography
-              variant="subtitle2"
-              component="small"
-              className={`${classes.helpText} ${inputs.createAsOrphan ? '' : classes.muted}`}
-            >
-              {formatMessage(messages.createAsOrphanHelpText)}
-            </Typography>
-          </Grid>
         )}
       </Grid>
     </form>
